make supressed and reported field in user class so both employer & jobseeker inherits it; remove unneeded job class since jobseeker replaces it;add better naming for columns of embedded contact class
