+++ src/main/java/usjobs/model/Contact.java
@@ -1,18 +1,24 @@
 package usjobs.model;
 
+import javax.persistence.Column;
 import javax.persistence.Embeddable;
 
 @Embeddable
 public class Contact{
-
+	
+	@Column(name="contact_first_name")
 	private String firstName;
 
+	@Column(name="contact_last_name")
 	private String lastName;
 
+	@Column(name="contact_title")
 	private String jobTitle;
 	
+	@Column(name="contact_phone")
 	private String phone;
 	
+	@Column(name="contact_email")
 	private String email;
 
 	public String getFirstName() {
